implemented (partially broken) control flow (if/else, while), switch/case coming later

+++ src/chunk.c
@@ -13,17 +13,17 @@ void initChunk(Chunk* chunk) {
 }
 
 void freeChunk(Chunk* chunk) {
-    FREE_ARRAY(uint32_t, chunk->code, chunk->capacity);
-    FREE_ARRAY(unsigned int, chunk->lines, chunk->capacity);
+    FREE_ARRAY(uint8_t, chunk->code, chunk->capacity);
+    FREE_ARRAY(int, chunk->lines, chunk->capacity);
     freeValueArray(&chunk->constants);
     initChunk(chunk);
 }
 
-void writeChunk(Chunk* chunk, uint32_t byte, int line) {
+void writeChunk(Chunk* chunk, uint8_t byte, int line) {
     if (chunk->capacity < (chunk->count + 1)) {
         int oldCapacity = chunk->capacity;
         chunk->capacity = GROW_CAPACITY(oldCapacity);
-        chunk->code = GROW_ARRAY(uint32_t, chunk->code,
+        chunk->code = GROW_ARRAY(uint8_t, chunk->code,
             oldCapacity, chunk->capacity);
         chunk->lines = GROW_ARRAY(int, chunk->lines,
             oldCapacity, chunk->capacity);
