+++ src/include/vm.h
@@ -6,7 +6,7 @@
 
 typedef struct {
     Chunk* chunk;
-    uint32_t* ip;
+    uint8_t* ip;
     Value* stack;
     int stackCount;
     int stackCapacity;
